+++ cmd/soroban-rpc/internal/methods/util.go
@@ -2,6 +2,7 @@ package methods
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/db"
@@ -12,13 +13,13 @@ func getProtocolVersion(
 	ledgerEntryReader db.LedgerEntryReader,
 	ledgerReader db.LedgerReader,
 ) (uint32, error) {
-	var protocolVersion uint32
 	readTx, err := ledgerEntryReader.NewCachedTx(ctx)
 	if err != nil {
 		return 0, err
 	}
 	defer func() {
-		_ = readTx.Done()
+		doneErr := readTx.Done()
+		err = errors.Join(err, doneErr)
 	}()
 
 	latestLedger, err := readTx.GetLatestLedgerSequence()
@@ -26,28 +27,16 @@ func getProtocolVersion(
 		return 0, err
 	}
 
-	_, protocolVersion, err = getBucketListSizeAndProtocolVersion(ctx, ledgerReader, latestLedger)
-	if err != nil {
-		return 0, err
-	}
-	return protocolVersion, nil
-}
-
-func getBucketListSizeAndProtocolVersion(
-	ctx context.Context,
-	ledgerReader db.LedgerReader,
-	latestLedger uint32,
-) (uint64, uint32, error) {
 	// obtain bucket size
 	closeMeta, ok, err := ledgerReader.GetLedger(ctx, latestLedger)
 	if err != nil {
-		return 0, 0, err
+		return 0, err
 	}
 	if !ok {
-		return 0, 0, fmt.Errorf("missing meta for latest ledger (%d)", latestLedger)
+		return 0, fmt.Errorf("missing meta for latest ledger (%d)", latestLedger)
 	}
 	if closeMeta.V != 1 {
-		return 0, 0, fmt.Errorf("latest ledger (%d) meta has unexpected verion (%d)", latestLedger, closeMeta.V)
+		return 0, fmt.Errorf("latest ledger (%d) meta has unexpected verion (%d)", latestLedger, closeMeta.V)
 	}
-	return uint64(closeMeta.V1.TotalByteSizeOfBucketList), uint32(closeMeta.V1.LedgerHeader.Header.LedgerVersion), nil
+	return uint32(closeMeta.V1.LedgerHeader.Header.LedgerVersion), nil
 }
